Move inference test fixtures and dedupe solution serialization (#51 review).
Shared fixtures live under tests/fixtures/; analytic solution wire helpers
are consolidated with a module-level eligibility import.
