+++ packages/api/tests/conftest.py
@@ -1,7 +1,6 @@
 """Shared pytest fixtures for API package tests."""
 
-from tests.test_military_score_inference_actions import (  # noqa: F401
-    _observation,
+from tests.fixtures.military_score_inference import (  # noqa: F401
     sample_turn,
     synthetic_catalog_context,
 )
